+++ frontend/src/components/JourneyEdit.js
@@ -9,8 +9,6 @@ import {
     Button,
     DatePicker,
     InputNumber,
-    TreeSelect,
-    Switch,
     Row,
     Col,
     Space,
@@ -20,25 +18,15 @@ import { MinusCircleOutlined, PlusOutlined } from '@ant-design/icons';
 import moment from 'moment';
 import SearchComplete from "./SearchComplete";
 import CustomSlider from "./CustomSlider";
+import JourneyMap from "./JourneyMap";
 const { Title } = Typography;
 
 
 function JourneyEdit() {
     const [ymaps, setYmaps] = useState(null);
     const [points, setPoints] = useState([]);
-    const [currPoints, setCurrPoints] = useState([]);
-    const [smallCost, setSmallCost] = useState(0);
-    const [avgCost, setAvgCost] = useState(0);
-    const [largeCost, setLargeCost] = useState(0);
-    const defmapState = {
-        center: [55.751574, 37.573856],
-        zoom: 5
-    };
 
-    const width = 600;
-    const height = 500;
-
-    function onMapClick(event) {
+    const onMapClick = (event) => {
         let position = event.get('coords');
         console.log(position);
         ymaps.geocode(position)
@@ -50,7 +38,8 @@ function JourneyEdit() {
             });
     }
 
-    function getPointGeoData(address) {
+    const onMapLoad = (ymaps) => {
+        setYmaps(ymaps);
     }
 
     const onAddressSelect = (index, value) => {
@@ -99,46 +88,6 @@ function JourneyEdit() {
             pointsList.push(geo);
         }
         pointsForm.setFieldsValue({points: pointsList});
-
-    }
-
-    const renderPoints = () => {
-        const ps = points.map((point, index) => {
-            return {geoData: point ? point.geoData : undefined, index: index};
-        }).filter(point => point.geoData);
-        return (
-            <>
-                {ps.map(el =>
-                    <Placemark
-                        key={el.index + 1}
-                        geometry={el.geoData.coordinates}
-                        properties={{
-                            hintContent: el.geoData.text,
-                            balloonContent: el.geoData.balloonContent,
-                            iconContent: el.index + 1
-                        }}
-                        options={{
-                            draggable: false
-                        }}
-                    />
-                )}
-                <Polyline
-                    key={1}
-                    geometry={
-                        ps.map(el =>
-                            el.geoData.coordinates
-                        )
-                    }
-                    options={{
-                        balloonCloseButton: false,
-                        strokeColor: '#1E90FF',
-                        strokeWidth: 4,
-                        strokeOpacity: 0.5,
-                    }}
-                />
-            </>
-        )
-
 
     }
 
@@ -258,19 +207,11 @@ function JourneyEdit() {
                         </Form>
                     </Col>
                     <Col>
-                        <Map
-                            defaultState={defmapState}
-                            width={width}
-                            height={height}
-                            onClick={onMapClick.bind(this)}
-                            onLoad={ymaps => {setYmaps(ymaps);console.log(ymaps)}}
-                        >
-                            <SearchControl
-                                options={{ provider: 'yandex#search' }}
-                            />
-                            {renderPoints()}
-                        </Map>
-
+                        <JourneyMap
+                            onMapLoad={onMapLoad}
+                            onMapClick={onMapClick}
+                            pointsList={(pointsForm.getFieldsValue("points").points || [])}
+                        />
                     </Col>
                 </Row>
             </YMaps>
